animeproj/pyproject.toml
[tool.poetry]
name = "animeproj"
version = "0.1.0"
description = "Это платформа для любителей аниме, предлагающая возможность обмена мнениями, создания избранных списков, а также поиск и группировку аниме по различным параметрам."
authors = ["DmitryAce <47420776+DmitryAce@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asgiref = "3.8.1"
build = "1.2.1"
cachecontrol = "0.14.0"
certifi = "2024.7.4"
charset-normalizer = "3.3.2"
cleo = "2.1.0"
colorama = "0.4.6"
crashtest = "0.4.1"
distlib = "0.3.8"
django = "5.1"
django-cleanup = "8.1.0"
djangorestframework = "3.15.2"
djangorestframework-simplejwt = "5.3.1"
drf-yasg = "1.21.7"
dulwich = "0.21.7"
fastjsonschema = "2.20.0"
filelock = "3.15.4"
idna = "3.8"
inflection = "0.5.1"
installer = "0.7.0"
jaraco-classes = "3.4.0"
keyring = "24.3.1"
more-itertools = "10.4.0"
msgpack = "1.0.8"
packaging = "24.1"
pexpect = "4.9.0"
pillow = "10.4.0"
pkginfo = "1.11.1"
platformdirs = "4.2.2"
poetry = "1.8.3"
poetry-core = "1.9.0"
poetry-plugin-export = "1.8.0"
ptyprocess = "0.7.0"
pyjwt = "2.9.0"
pyproject-hooks = "1.1.0"
pytz = "2024.1"
pywin32-ctypes = "0.2.3"
pyyaml = "6.0.2"
rapidfuzz = "3.9.6"
requests = "2.32.3"
requests-toolbelt = "1.0.0"
setuptools = "73.0.1"
shellingham = "1.5.4"
sqlparse = "0.5.1"
tomlkit = "0.13.2"
trove-classifiers = "2024.7.2"
tzdata = "2024.1"
uritemplate = "4.1.1"
urllib3 = "2.2.2"
virtualenv = "20.26.3"
wheel = "0.44.0"
django-apscheduler = "^0.6.2"
django-restframework = "^0.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
